1. Use unsigned int for the magic's type. 2.Region-wide mem-copying in Clone() changed to field-by-field. 3. Checking magic number replaced by a special function s_InfoIsValid(const SConnNetInfo* info). JIRA: CXX-7869
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@73494 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/connect/ncbi_connutil.h
@@ -188,7 +188,7 @@ typedef struct {  /* NCBI_FAKE_WARNING: ICC */
     const char*     http_user_header; /* user header to add to HTTP request  */
     const char*     http_referer;     /* default referrer (when not spec'd)  */
     NCBI_CRED       credentials;      /* connection credentials (optional)   */
-    unsigned short  magic;            /* to detect version skew              */
+    unsigned        magic;            /* to detect version skew              */
 
     /* the following field(s) are for the internal use only -- don't touch!  */
     STimeout        tmo;              /* default storage for finite timeout  */
